requirements.txt
flask>=2.3.2
requests==2.25.0
MarkupSafe>=2.1.2
json2html==1.3.0
plotly>=5.14.1
pandas>=1.5.1
python-dotenv==0.21.0
PyMySQL==1.0.3
mysqlclient==2.1.1